+++ trunk/src/faust-cc/balance.cc
@@ -11,19 +11,19 @@ void init(int samplingFreq)
 
 void compute(int count, float *input0, float *input1, float *output0, float *output1)
 {
-	double 	fSlow0 = fslider0;
-	double 	fSlow1 = (1 - max(0, fSlow0));
-	double 	fSlow2 = (1 - max(0, (0 - fSlow0)));
+	float 	fSlow0 = fslider0;
+	float 	fSlow1 = (1 - max(0, fSlow0));
+	float 	fSlow2 = (1 - max(0, (0 - fSlow0)));
 	for (int i=0; i<count; i++) {
-		output0[i] = (FAUSTFLOAT)(fSlow1 * (double)input0[i]);
-		output1[i] = (FAUSTFLOAT)(fSlow2 * (double)input1[i]);
+		output0[i] = (FAUSTFLOAT)(fSlow1 * (float)input0[i]);
+		output1[i] = (FAUSTFLOAT)(fSlow2 * (float)input1[i]);
 	}
 }
 
 static struct RegisterParams { RegisterParams(); } RegisterParams;
 RegisterParams::RegisterParams()
 {
-	registerVar("amp.balance","Balance","S","",&fslider0, 0.0, -1.0, 1.0, 0.1);
+	registerVar("amp.balance","Balance","S","",&fslider0, 0.0f, -1.0f, 1.0f, 0.1f);
 	registerInit(init);
 }
